+++ src/Faithlife.Build/AppRunner.cs
@@ -19,5 +19,22 @@ namespace Faithlife.Build
 		/// <param name="args">The arguments to send to the command-line app.</param>
 		public static void RunApp(string path, params string[] args) =>
 			Command.Run(path, ArgumentEscaper.EscapeAndConcatenate(args.Where(x => x != null)));
+
+		/// <summary>
+		/// Runs the specified .NET Framework command-line app.
+		/// </summary>
+		/// <param name="path">The path of the command-line app.</param>
+		/// <param name="args">The arguments to send to the command-line app.</param>
+		/// <remarks>On Linux and macOS, Mono is used to run the app.</remarks>
+		private static void RunDotNetFrameworkApp(string path, params string[] args)
+		{
+			if (BuildEnvironment.IsUnix())
+			{
+				args = new[] { path }.Concat(args).ToArray();
+				path = "mono";
+			}
+
+			RunApp(path, args);
+		}
 	}
 }